scripts/etl/requirements.txt
psycopg2-binary>=2.9.9
geopandas>=0.14.0
pandas>=2.0.0
shapely>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
